requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.28
flask-jwt-extended
reportla
python-dotenv